feat: added env based installation for unsupervise setups

+++ script/setup/ubuntu/misc/neovim.sh
@@ -4,14 +4,20 @@
 sudo apt-get install neovim -y
 
 # setup neovim configs (if they want)
-read -p "install nvchad? (Y/n) " IFS_YES
-IFS_YES="${IFS_YES:-Y}"
+IFS_YES="${NVCHAD:-}"
+if [ -z "$IFS_YES" ]; then
+    read -p "install nvchad? (Y/n) " IFS_YES
+    IFS_YES="${IFS_YES:-Y}"
+fi
 if [[ $IFS_YES =~ ^[Yy]$ ]]; then
     echo "adding nvchad..."
     git clone git@github.com:onixldlc/nvim-config.git $HOME/.config/nvim
 
-    read -p "install nvchad to root? (Y/n) " IFS_YES2
-    IFS_YES2="${IFS_YES2:-Y}"
+    IFS_YES2="${NVCHAD_ROOT:-}"
+    if [ -z "$IFS_YES2" ]; then
+        read -p "install nvchad to root? (Y/n) " IFS_YES2
+        IFS_YES2="${IFS_YES2:-Y}"
+    fi
     if [[ $IFS_YES2 =~ ^[Yy]$ ]]; then
         echo "adding nvchad to root..."
         sudo rm -d /root/.config/nvim
